requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
PyJWT==2.6.0
shortuuid==1.0.11
SQLAlchemy==1.4.45
SQLAlchemy-serializer==1.4.1
Werkzeug==2.2.2
zipp==3.11.0
